d'ou c'était du long ??????????

+++ good_everywhere/normal.py
@@ -90,7 +90,7 @@ for i in range(nbbatchs):
     optimizer.step()
 
     with torch.no_grad():
-        z = (z[:, 1, :, :] > z[:, 0, :, :]).long()
+        z = (z[:, 1, :, :] > z[:, 0, :, :]).float()
         for j in range(batchsize):
             stats[batchchoise[j]][0] += torch.sum((z[j] == y[j]).float() * D[j])
             stats[batchchoise[j]][1] += torch.sum(D[j])
